now there's no while(true)!

+++ ConsoleApp1/Application/IDataBase.cs
@@ -6,7 +6,7 @@ namespace Organizer
 {
     public interface IDataBase
     {
-        public Byte[] GetData(long userId, int dataTypeId, int dataSubTypeId, DateTime time);
-        public void SaveData(long userId, int dataTypeId, int dataSubTypeId, DateTime time, Byte[] data);
+        public Byte[] GetData(int userId, int dataTypeId, int dataSubTypeId, DateTime time);
+        public void SaveData(int userId, int dataTypeId, int dataSubTypeId, DateTime time, Byte[] data);
     }
 }
